Abstract Rfc logic to AbstractRfcFilter
