+++ tests/Unit/Shared/RfcFilterImplementation.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace PhpCfdi\SatWsDescargaMasiva\Tests\Unit\Shared;
+
+use PhpCfdi\SatWsDescargaMasiva\Shared\AbstractRfcFilter;
+
+final class RfcFilterImplementation extends AbstractRfcFilter
+{
+}
